change int to double

+++ Codecool.MarsExploration.MapExplorer/MarsRover/MarsRover.cs
@@ -7,5 +7,5 @@ using System.Threading.Tasks;
 
 namespace Codecool.MarsExploration.MapExplorer.MarsRover
 {
-    public record MarsRover(int Id, Coordinate currentPosition, int SightDistance, IEnumerable<Coordinate> Coordinates); 
+    public record MarsRover(int Id, Coordinate currentPosition, double SightDistance, IEnumerable<Coordinate> Coordinates); 
 }
